LCDCombo: change focus policy to Qt::NoFocus
previously, we used `Qt::ClickFocus`. It seems this was done at the same time when also introducing this setting to `LCDSpinBox`. In the later, it is quite useful since it prevents the widget to be reachable via tab (this should be used for the editors exclusively) while still allowing it to be modified using keyboard event after clicking it.

But for `LCDCombo` there is no advantange. Clicking the widget will open a popup menu and regardless of the focus policy of the widget itself, the user can still navigate the popup using keyboard. Instead, it has the negative side effect that the widget swallows all keyboard event (including shortcuts) after usage and the user has to focus out by clicking somewhere else first in order to start/pause transport via keyboard again.