EventQueue: rework using deque
since the Windows pipeline fails the EventQueue test after the refactoring anyway, this is probably a good point in time to simplify the `EventQueue` itself.

Instead of having a fixed sized array with volatile read and write indices, we now use a `deque` pushing new events to the back and popping them from the front. Also, instead of the `Event::Type::None`, we now just use a `nullptr` to indicate that there is no event present at the momemnt