TransportPosition: avoid redundant pattern add
When assigning the content of one `TransportPosition` to another one, there is no need to add the virtual patterns contained in the playing patterns individually. They are already present in the `TransportPosition` provided as argument.

Due to the checks in `PatternList::add` this did not caused harm but just resulted in a false positive info log message