PatternEditor: drag edit either length or property
doing both at the same time reults in a horrible UX.

Recently I had to use the drag feature to introcude user-defined note lengths. But it was horrible. The slighest vertical movement caused the properties of the note to be overwritten.

Now, whether the initial drag movement is vertical or horizontal decides which property will be edited