PatternEditor: indicate disabled quantization
every time the user presses the <kbd>Alt</kbd> modifier during a move or copy operation, the selected note(s) can be moved off grid. While this is in line with the more fine-grained value control provided by this modifier within the `NotePropertiesRuler`, it still feels like a hidden feature.

In order to better integrate it in terms of UI/UX, there is now a visual representation of "quantization disabled": all additional possible ticks will be draw with a dotted grid line. The regular grid lines corresponding to the current resolution are still rendered as solid lines and are more pronounced compared to the additional ones.