PatternEditor: disable quant. using quant. button
previously, the quantization button did only disabled quantization for recorded notes but neither for those inserted via keyboard or mouse nor those interacted with those inputs. This felt a little strange and made it hard to determine effect of the quantization button (in the pattern editor panel) by toying with the interface.

Now, unchecking the button does also disable quantization in the overall pattern editor and results the corresponding visualization.

In addition, I moved the quantization state into `PatternEditorPanel` and ensured it is updated on all relevant events (its side effects should affect both the visible editor and note ruler)

The default value of the quantization is "on". Users need to explicitly disable it (or have to did this at some point in the past) to see this feature. I do not think it will confuse established users