PatternEditor: do not update modifiers on release
of both mouse button and keyboard. This way the last state rendered will be applied.

Maybe it is something personal but a couple of times I was slightly frustated because I had - again - a little race condition between modifier and keyboard key or mouse button. I release the modifier slightly earlier and the note moved to a place I did not intended or was moved instead of copied.