PatternEditorPanel: requantize on moveCursorRight
Previously, when moving the cursor to the right-most possible tick, turning off quantization and moving the cursor even further right, the cursor stays at the resulting position when reenabling quantization and hitting the right arrow key again.

This itself felt quite natural. But this did not coincides with how sihouettes of moved notes are handled. They do depend on the offset provided by `PatternEditor::movingGridOffset` and do depend on whether quantization is active or not.

To make both act the same, the cursor is now re-quantized when hitting the right arrow key in the scenario above. Now that we feature a visual feedback for quantization being turned off, this feels better than the previous approach