PatternEditorSidebar: taking hearNotes into account
I decided to take `Preferences::getHearNewNotes` into account in the `PatternEditorSidebar` as well.

The whole feature of previewing notes of an instrument by clicking the sidebar feels like some legacy feature since the same thing can be done even better within the `InstrumentEditor`. (Though, when using multiple components this is not true anymore. But I doubt any user has yet harness this particular feature just for multiple components).

Experienced users know they need to click in the left-most part of the sidebar when switching to another row - e.g. in order to access different properties in the ruler - without making a sound - e.g. when programming while playback is rolling. But this is an anti pattern and users dealing with the application only sporadically (this includes me) will often have to relearn this trick, which might be frustrating.

Instead, the "hear notes" button in the `PatternEditorPanel` does now affect notes triggered via the sidebar as well. When unchecked, the whole pattern editor will be muted. Otherwise, both note insertion and clicking the sidebar will make sounds