PatternEditorPanel: cache patterns to show
with pattern editor being locked and fast forwarding/backwarding to different columns, I had some note misses in the `PatternEditorPanel::m_db` in `PatternEditor::drawNote`. Turns out there was a small delay between the relocation event and the playing patterns changed event (rounding issue due to the JACK server. Something that is expected) causing a DB update on the old set of playing patterns followed by a rendering using the newer set.

This itself is nothing harmful since the latter playing patterns changed event would cause the DB to be refreshed and the editors to be rerendered. But it is not clean and causes some error messages.

Instead, we now cache all patterns to show instead of always retrieving the raw ones from the core and update this set on the corresponding events