PatternList: fix pattern add
in `TransportPosition::set` we do not add virtual patterns anymore since at this point they have already been added to the `PatternList`s and we just need to added all individual patterns.

But there was still a ?weird? sanity check in place prohibiting adding a pattern in case another one contained it as virtual one.