MAX_NOTES -> H2Core::nTicksPerQuarter
The option `MAX_NOTES` in `CMakeLists` has been dropped and replaced by (4 *) `H2Core::nTicksPerQuarter`.

This new name makes it more evident what the variable is actually representing (I had to look this up a couple of times already). But even more importantly by refactoring it from a user option into a constexpr we do not pretend anymore that it can actually be changed seamlessly. Both the `PatternEditor` grid and the column calculation are not able to deal with other values. And since no one complaint about this in the last 11 (!) years it seems no one ever altered it.