SMF: derive ticks per quarter
for some reason the resolution in MIDI files (ticks per quarter) is for times as high as in the `AudioEngine`. This has been since 2008 and there is no hint in the git history why.

To make this factor more evident, `SMF::nTickFactor` was introduced and `SMF::nTicksPerQuarter` is obtained as a product of it and `H2Core::nTicksPerQuarter`.

In the current state of the code there is no reason at all to keep this more fine-grained resolution. But I intend to introduce humanization into the MIDI export at some point soon. And for this it will be an advantage