SMF: use zero based ticks in events
the tick position starting at `1` was not incorrect. But it was still very weird as the position of the notes corresponding to the NoteON and NoteOFF events was off by `1` since the tick position in the `AudioEngine` is zero based.

The tick in the SMF and events have been aligned and are now zero based too. Since only relative values were used in the code, this does not make a difference in terms of the resulting MIDI files