Revert "SMF: drop custom sortEvent in favor for std::sort"
The sort algorithm of the std library produces inconsistent result on Linux and macOS. Occassionally, notes appearing at the same time are ordered differntly (on macOS: the one with the highest pitch first, on Linux and Windows: the other way around).

Instead, of further tweaking the algorithm for consistent results, we just revert to the original home brewed one.

This reverts commit 764a2400e5fa62693816b1c952a8e3a9bebdec66.