SMF: track-based SMFWriter classes
instead of maintaining only `EventList`s in the three `SMFWriter` classes, each one does now contain all `SMFTrack`s as well. Those tracks are only added at a later point when the song was already traversed by the writer. The `save` method does not add `SMFEvent`s to the EventList directly anymore, but calls the `addEvent` method and lets the writer handle it internally.

In addition, the `SMF1WriterMulti` does now use a `map` instead of a `vector` to store its per-instrument tracks/event lists. The instrument ID is used as map key which is a lot more resilient against drumkit changes.

Sorting of events was moved into `SMFTrack` and does now order meta events first (in case they appear at the same tick as note on and off events).

Now all MIDI unit tests artifacts are identical and we are set up to support meta events introduced by the `Timeline` in the next step