ComponentsEditor: accommodate scrollbar
now that we can have multiple `ComponentView`s expanded at the same time, we also need to account for the scrollbar of the overall `ComponentsEditor`. And that one was quite a pain.

Since the scrollbar of Qt is quite big (in terms of width), it would cover some widgets and would not look nice at all when just displaying it as needed. I also tried to dynamically change the size of the `InstrumentRack` whenever the scrollbar became visible. But this size change felt even worse.

I ended up increasing the width of `InstrumentRack`, `InstrumentEditorPanel`, `InstrumentEditor`, `SoundLibraryPanel`, and `ComponentsEditor` by the width of the scrollbar. `ComponentView` keeps its old width and the scrollbar is always visible. The latter is not visually pleasing but at least it yields a strong hint that one can have multiple components