InstrumentEditor: increase width by 12 pixel
what a pain! These pixel-based graphics have definitely were out their welcome!

In order to account for the additional size due to the scroll bar in `ComponentsEditor`, the `InstrumentEditor` pixel graphic was increased to cover 12 additional pixels, all affected widgets have been moved to the corresponding places, and some overall fine tuning of the widget position was done to make everything work with this new width