Theme: add dedicated mute and solo colors
historically the checked solo button was highlighted using `ColorTheme::m_accentColor`. Not an ideal solution, but ok. But lately we want to color elements, like soloed layers in `LayerPreview`, to emphasize that they are soloed. This can not be done using the accent color, since that one is already used all over the place. It would not be evident what is going on.

Instead, a dedicated solo color is introduced and exposed in the color theme within the `PreferencesDialog`. And while at it, I also introduced a dedicated mute color.

`ColorTheme::m_buttonRedColor` is kept since there are a couple of use cases where the red color is used to indicate errors instead of a muted state. It would be nice to distinct them visually too