Instrument: moving sampleSelectionAlgo into `InstrumentComponent`
there was another flaw in the drumkit definition: the algorithm using which the final sample will be selected is displayed in the (former "Layers") "Components" tab of the instrument editor but stored in `Instrument`. This means there is only a single value affecting all `InstrumentComponents` the same way.

This is neither right from the UI/UX perspective nor from the functional one. Depending on which samples are combined into a component, they could very well use different selection algorithms.

Since the drumkit definition will already be altered for version 2.0 and we did not officially publish the previously altered version, we can just fix this directly.

In addition, `SampleSelectionAlgo` as changed from an enum into the enum class `Selection`