DrumkitPropertiesDialog: dedicated save button and title
for editing the current drumkit.

When editing the current drumkit, which is part of the current song, things are quickly getting confusing since the dialog looks exactly like the one used for editing drumkit properties in the SoundLibrary.

To distinguish those variants, the window title does now indicate whether the current drumkit is edited and save button displays either "Save drumkit" or "Save song"