Note: no instrumentID reset in mapTo
there was a flaw in the `mapTo` method mapping the note to drumkit.

In the previous version we did reset the instrument ID to `-1` (`EMPTY_INSTR_ID`) in case the note does both hold a type and is mapped to an instrument without type not featuring the instrument id stored in the note.

While this is sufficient to map from a kit with types to one without and back without any loss of information, it has severe limitations. When starting with a big kit with types, all notes will have both types and instrument IDs. Now, the note/pattern is mapped to a kit without types and only a subset X of instruments. All notes of superfluous instruments would get an instrument ID of `-1`. Now, when mapping to a kit without types of size X+Y there would not be any note in the last Y instruments because all possible candidates have been reset priorly. When mapping directly from the original kit to the one of size X+Y, however, at least some of the Y instrument would bear notes. That is not ok.

Instead, instrument IDs will not be reset. Whether a note is mapped to a drumkit or not is now solely indicated by its `m_pInstrument` being `nullptr` or a valid `shared_ptr<Instrument>`.

While this sounds like a smaller change, this has huge consequences on the design of the pattern editor