NoteTest: add unit test for mapping to legacy kits
honestly, I do not know why I did not implemented the whole mapping feature in a test-driven way the begin with. Would have saved some time.

In `src/tests/data/pattern/noteTestTypeMisses.h2pattern` a note was changed to hold an instrument ID outside of the range of the legacy GMKit used in the new unit test