PatternEditor: supply type and id of note
instead of the corresponding row.

In edge-cases (in the new design, which does not reset the instrument ID of an unmapped note anymore) we could have multiple notes holding the same type but different instrument IDs. For a more concise presentation we merge all those notes into a single row of the pattern editor. But when performing an action on those notes we have to ensure we use type and instrument id stored in them directly. Else we might face loss of information