fix Drumkit::Context initialization
setting `Context::Song` in `Song::setDrumkit` to the loaded kit is too late. There might be some song-specific stuffed applied to the kit beforehand. Instead, `Context::Song` is now the new default value and we take care to apply it as early as possible when loading a kit from `SoundLibraryDatabase`