Note: more strict mapTo
the automatic mapping from typed instruments to non-typed ones did caused a lot of confusion. I tried an approach where the context `mapTo` was called decided about how the mapping should be done. But things just got more complex.

In the end I decided to simpligy the mapping instead. Now, notes with types are _only_ mapped to instrument with types and notes without types are mapped by order or id to instruments without id with the one exception of assigning an initial type to an instrument which was previously without type