Pattern: use DrumkitMap of song kit
when loaded as part of the song.

E.g. the user loads the GMRockKit, adds another instrument, and assigns a type. For this instrument she adds some patterns. Now, when saving, closing, and loading the song, the pattern has be loaded/mapped to the drumkit map of the modified GMRockKit version of the song. The name of the kit has not changed yet and the fallback is resort to the .h2map file of the original GMRockKit. Well, fallback should not be should not be triggered since the corresponding note does hold the type a well, but lets be on the safe side