Pattern: commonize type fallbacks on load
previously, missing types were only assigned when loading a recent pattern. But it is vital to do this in `Legacy::loadPattern` too.

The overall fallback has been redesigned. Instead of relying on the shipped drumkit maps exclusively, the fallback will now search all drumkits in the `SoundLibraryDatabase` first for one matching the provided name