Preferences: add option to enforceInstrumentName
in the naming convention of the JACK per-track output ports.

If set, the previous naming scheme using the track number and instrument name (cause the latter is not unique) is used. But this requires renaming ports each time the drumkit is switched. This can mess up existing connections. If not set, the instrument type will be used instead.