JackAudioDriver: truncate port name on max size
the JACK API specifies the port names (including the client name as a prefix) to not exceed a given maximum number of characters. We will honor this by truncating the port name in such a way there is still room for deduplication and appending the `_L` or `_R` channel suffix.

HOWEVER, this patch does not work on my machine. JACK is reporting a maximum number of `321` characters using `jack_port_name_size()`. But this is seemingly wrong and the true value has to be `256`. But this is not a Hydrogen related problem but a bug in JACK itself. I will include this patch regardlessly since it may be fixed in future JACK versions and it does not doe any harm