JackAudioDriver: use transient port names for death row
We have to ensure ports which are soon to be removed do not conflict with new ports being registered.

E.g. when enforcing instrument names as port names and having the first instrument with an ID != `0` called "New Instrument" and then replacing the kit with a new one, the only instrument of the new kit would hold the same name but a different id.

Instead, whenever a port is marked for death, we rename it by appending the suffix "_removed". This probably also make more clear what is happening with the port