JackAudioDriver: clean up death row before making ports
currently the cleanup the death row is only done automatically when stopping transport. This is itself a good idea in order to be real time friendly. But existing ports of instruments already rendered to clash with names of new ones. That is why we need to clean up the death row ports before attempting to create new ones