Parser: add options to en/disable ANSI colors in log
Previously ANSI colors in log messages had been disabled on Windows and enabled on all other platforms.

But in some edge cases this behavior was not that satisfying - e.g. writing log messages to a file on Windows and using an ANSI-capable editor (me) or doing the same using an editor being not ANSI-capable on other platforms (see #2136).

The default behavior did not change. But using this patch the user is able to overwrite in both cases.

Note that the CLI variable was not introduced into `h2cli`, `h2player`, and our test binary on purpose. The latter is only used internally and not shipped. The former two are not part of our macOS and Windows releases and are probably unknown to most users. Plus, issue coverage hints that they are used only scarsely. There is thus no advantage of allowing ANSI colors on Windows and negligible benefit for disabling them on other platforms