windows: replace -32bit option by -qt5
The `-32bit` option was dropped since MSYS2 has dropped that many packages that we can not use it anymore to build Hydrogen with.

However, the CMake `-DWIN64` option is not dropped yet. People may want to build it locally in another way.

Instead, `-qt5` is now introduced to build Hydrogen using Qt5 on Windows. Per default, Qt6 will be used.