HydrogenApp: fix event listener update
previously adding and removing `EventListener` in `HydrogenApp` was decoupled from their handling in `HydrogenApp::onEventQueueTimer`. This can quite some troubles since due to the handling of individual events listener can be removed or added. `onEventQueueTimer` accesses each listener via its index in a `std::vector`. On `Event::Type::UpdateSong` various listeners will be added/removed. In one local test case the number dropped from 169 to 123 causing the loop over all listener to skip a couple of them - like the `PatternEditorPanel` and `PlayerControl`. This resulted in visual artifacts.

Instead, we now use two queues in which we enroll all listeners to be added or removed. The actual update of the vector of active listeners will be done on the next event encountered.