BaseEditor: first stub
Starting the implementation of MADR 0007 by moving some parts of `PatternEditor` in a template parent class called `BaseEditor`.

It does not use a header - implementation file separation since the compiler needs to be aware of the particular instantiations required.

For some methods only the child class can implement purely virtual methods would be more fitting. But the whole point of `BaseEditor` is share common logic and member variables. AFAIK this is nothing an Abstract class can provide in C++