EditorBase:: do not update hovered elem in updateEditor
this function should only be used for redrawing as it very well can be called for all editors in the pattern editor section in a row. Updating the currently hovered notes on each call is not a good idea.

This might result in some glitches. When all required code is moved into `Editor::Base`, a cleanup and overhaul of the editor interface will be done addressing all these glitches