PatternEditor: move keyboard handling into Editor::Base
upon many things that have changed, the user is now able to created and remove notes in the `NotePropertiesRuler` by hitting enter or delete. It felt strange at first, I can not think of a single reason not to do it since the overall goal is to align the behavior of the different editors.