Move drag/draw into Editor::Base
`SelectionWidget::mouseDrag*Event` was renamed into `SelectionWidget::mouseDraw*Event` since by now it is only used to drawing properties or elements.

Right-click drag editing of note properties in `DrumPatternEditor` and `PianoRollEditor` has been dropped. Supporting this one together with right-click drawing would result in a bad UX. Instead, there will be an "edit" button next to the "draw" button using the the user can toggle to left-click drag a note instead of starting a selection. This will be another step towards a touch-compatible UX in the pattern editor