PatternEditor: support drawing in `DrumPatternEditor` and `PianoRollEditor`
using right-click drag or when activating draw mode, the user can now draw notes in our two main pattern editors similar to how it is supported in `SongEditor` for drawing patterns. The main intention behind this feature is to get a more consistent editor interface throughout Hydrogen. But I have to admit I was longing for this feature myself for quite a while :D