FileDialog: no native windows on Linux (#2165)
Occassionally users reported freezing of Hydrogen when interactive with native
file dialogs on Linux, e.g. #2165. Qts bug tracker is full of such tickets and
it seems to be a problem with the interaction of the Qt library with the
underlying windows manager. Since I (@theGreatWhiteShark) was never able to
reproduce this issue and we have no control of which Qt version will be combined
using with window manager, we will circumvent this issue by avoid all native
file dialogs.