PlayerControl: port to layout-based design
instead of loading pixmap backgrounds and moving the individual widgets to a fixed pixel position, `PlayerControl` is now based on a scalable layout design (as intended by Qt). This is way more easy to maintain.

Some changes have been done along the way
- the LEDs next to song and pattern mode buttons have been dropped
- the BeatCounter activation button was dropped. It will be replaced by a right-click context menu allowing to show and hide it soon.
- some of the vertical stacked widgets have been replaced by horizontal ones. In the upcoming commits the whole toolbar will be redesigned in a more or less horizontal one.

Some beat counter related code was commented. It will be fixed and reintroduced in the next commits.