PlayerControl: introduce visibility buttons
there might be a historical reason why only the visibility of the mixer and instrument rack can be toggled via buttons in the `PlayerControl`. But as of 2025 we should either consistently expose the visibility of (important) dialogs/widgets or none. I would opt for the first one for maximum backward compatibility with the option to hide them via the right-click context menu (not implemented yet).

All those buttons to feature an uniform size and hold a dummy character as label for now. It is intended to supply each one of them a SVG icon once their final size is (more or less) fixed.