drop CpuLoadWidget
instead of the `CpuLoadWidget` in `PlayerControl`, we do now feature a textual representation of the CPU load in `Footer`.

This was done to a allow for a pure horizontal and button-focused design of the `PlayerControl` as well as to get a more modern design. The colored bar did not looked that nice anymore and is far less precise than the textual representation