Footer: implement XRun label
previously, `CpuLoadWidget` was used to indicate XRun events by coloring its boundaries red for about 2 seconds. Now that the widget is gone, we have a numerical display in `Footer` showing the number of XRun events which occurred since the last driver start. If the amount is `0`, the label will be hidden