PlayerControl: fuse MetronomeLED into MetronomeButton
`MetronomeLED` within `PlayerControl` was dropped and, instead, `MetronomeButton` is now a new class/file which changes its background color according to the current metronome tick. This was done for two reasons: first, to allow for a purely horizontal `PlayerControl` and, second, because the former `MetronomeLED` was just plain ugly.

In addition, we do not use a constant flash/highlight time on a metronome event but an adoptive one always set to half of the length of one beat