PlayerControl: starting port to QToolBar
I almost implemented a button providing a dropdown menu whenever `PlayerControl` is not wide enough to display all its widgets. But luckily I realized that Qt is already offering such a widget in `QToolBar`. Apart from that this class is the ideal base of our current `PlayerControl`.

It is, again, a lot of refactoring and restyling but dropping our custom `PlayerControl` for `QToolBar` will make Hydrogen more easy to maintain and it allows it to age with more dignity.

All buttons will be replaced with actions, but bpm spinbox, time display, and `BeatCounter` will stay as is