PatternEditorPanel: no tool bar in sidebar
I am not 100% sure about this yet. Final adjustments will be made once the toolbar of the `SongEditorPanel` are ported as well.

But for now omitting the sidebar feels right. It stresses that it applies to the pattern editors and the sidebar is concerned with entirely different things: an interface with the drumkit. Apart from that, confiding the buttons in the sidebar area by manually tuning pixel sizes and spaces is tedious and does not scale well