BeatCounter: more descriptive Pref and TapTempo action
some configurations in `Preferences` concerning the `BeatCounter` have been refactored into more descriptive enum classes.

The Tap Tempo feature is now accessible via a third option in the tap button of `BeatCounter`. It uses the same icon as the tap version of the beat counter but with all other buttons and labels disabled. This will also make the feature of hiding the whole `BeatCounter` obsolete