BpmTap: move tap button to the left & hide on tap tempo
the tap button has been moved to the left. This way it a) integrates better visually with the bpm spinbox and b) the button does not move anymore when changing its associated action.

In addition, now all +/- buttons and labels are hidden when tap tempo is selected. This should make even more evident what is going on