MidiControlButton: replace LEDs by SVGs
on each event (as well as on deactivation in future commits) the color of the SVG will be altered to indicate the state change.

This does integrate much nicer into the overall `MainToolBar` as well as is probably easy to handle in the future. Instead of requiring details Qt and C++ knowledge to change the widget (when rendering it from scratch), a designer can improve the SVG separately.

For now I just use some placeholder SVGs