MidiControlButton: increase highlight timeout
Previously, it was more easy to distinguish individual MIDI events. But the icons flicked quite a lot and it was rather annoying. In the end, those icons are meant to indicate whether MIDI input/output is working. More details about individual events will be accessible from the dialog opening when clicking the button