MidiInput: drop m_bActive
this member was present since the beginning of the git history and only used at a single point. There - in `AlsaMidiDriver` it checked in `midi_action` whether the driver was already active. But since the `AudioEngine` sets the driver active immediately after calling `MidiInput::open` and `midi_action` can only be called after the associated pthread was started in `open`, this check is redundant.

It is probably some sort of relic from prehistoric time there was still no decent way to check the state of the `AudioEngine` and meant to wait until its ready