MidiDrivers: add isInputActive() and isOutputActive()
if either is not the case, the corresponding icon in the main tool bar will be deactivated.

But this state is only partially helpful as if only informs the user about the overall capability of being connected with a MIDI input or output device. Since those connections can be established outside of Hydrogen without our notice, there is no reliable way to indicate an actual connection using those icons.