AppVeyor/Windows: reinstall some basic packages
Lately our Windows build pipeline does not work due a version conflict

> c:\msys64\usr\bin\pacman --noconfirm -S -q mingw64/mingw-w64-x86_64-libarchive
resolving dependencies...
looking for conflicting packages...
error: failed to prepare transaction (could not satisfy dependencies)
:: installing mingw-w64-x86_64-gcc-libs (15.1.0-7) breaks dependency "mingw-w64-x86_64-gcc-libs=14.2.0-2" required by mingw-w64-x86_64-gcc

this indicates that the `pacman` cache is out of date and the versions installed are not compatible with those we want to install. An upgrade of all package can not be done as this would update `pacman` itself and force the console instance to quit. Instead, we tear out a basic package and all dependent packages as well as all dependencies no longer required. Afterwards, we manually reinstall everything that was removed.

This sufficiently upgrade our installed package base, but it requires a lot of time. Maybe the AppVeyor people will update their system at some point in the future. But I suspect not. Maybe this is done on purpose or they prepare to drop the image entirely.