Sampler: bypass solo checks for preview instrument
the preview instrument is meant to render audio unrelated to the current song or kit (samples no longer present in `SampleEditor`, instruments of others kits in `SoundLibraryPanel`, and audio files in `AudioFileBrowser`). So, it should not be affected by any mute/solo settings of the current kit either. Instead, it bypasses everything except the state of the master mute button