Filesystem: fix sample path preparation (#2174)
a couple of things went wrong while handling the sample paths. While adding support for Qt6 a couple of things had to be adjusted in the corresponding methods leading to a regression due to the `QString::right` method being used incorrectly. This method is not designed to return the string right of a specific index n but to return the n right most characters. This lead to the drumkit name being detected incorrectly.

In releases <= 1.2.4 took the path `/<path_to_drumkit_folder>/<drumkit_name>/<sample_filename>` and checked whether `<drumkit_name>` is actually contained in the corresponding drumkit folder. This section was ported incorrectly into 1.2.5 and is dropped now. Since the sample paths do also need preparation when saving them the first time, there is no advantage in checking for the drumkits existence. If a proper check for existence would be required, the whole sample path could be checked directly right away.

Another thing that went unnotice in all prior release was that duplicated file separators, like `/<path_to_drumkit_folder>/<drumkit_name>//<sample_filename>`, would mess up the handling as well. It is hard to tell whether this took place in productive systems, but the method was hardened to avoid this problem.

The reason all of this was not detected prior to the 1.2.5 release by either manual testing or our unit tests, was that Hydrogen did fall back to use the absolute sample file path instead. When staying on the same device, everything will just run smoothly. But the resulting .h2song files will not be portable anymore.

The problem got uncovered due to a second minor bug causing the AppImage version of Hydrogen include its mount point when storing absolute paths.