SoundLibraryDatabase: add installed kits in AppImage too
while debugging it proved a little bit inconvenient that the AppImage version of Hydrogen does not list the drumkits installed on the system. Since it ships its own system drumkit folder, it is not totally wrong. But we could still list all kits installed as session kits. Otherwise, the user would need to load them manually into the session. And especially for new users installing the debian package containing a bunch of kits this might be too difficult