PatternEditor: fix eventPointToColumnRow
previously two main way to retrieve a column were provided by this function: as `m_nColumn` (quantized or not) and `m_nRealColumn` (never quantized). Even with quantization off / at lowest resolution those two still were not identical. `m_nColumn` is always rounded to the next nearest grid cell while `m_nRealColumn` is floored. At the default resolution and zoom level the differences were almost unnoticable. But when zooming in all the way, the `m_nRealColumn` approach felt like a bug.

Instead, we do now _always_ round to the next nearest grid cell and the only additional parameter provided to `eventPointToColumnRow` is now whether or not the quantization should be honored. In this form we can finally commonize it