PatternEditor: allow to create notes in NotePropertiesRuler
it is already possible to create notes in `NotePropertiesRuler` by pressing <kbd>Enter</kbd> and it feels awkward not to do so using mouse click too. Well, one has to get used to it a bit. But I think UX benefits a lot. This way one can easily create new notes in a specific property. No need to create a default one in `DrumPatternEditor` or `PianoRollEditor` first and than to switch back to `NotePropertiesRuler` just to adjust the property.

Still, drawing in `NotePropertiesRuler` should behave differently and should not create notes.