MainToolBar: move editor mode buttons to the left
the buttons to switch between pattern and song mode have been moved to the left of the time display and to the right of the input mode group. This way all editor-related things are grouped together.

For those four leftmost item `QAction` and `QActionGroup` is now used instead of `QButtonGroup`