SongEditor: fix highlight rendering
highlighting of partially drawn grid cells was off. In addition, anti aliasing is now used activated on highlighting. This ensure the highlight is uniformly drawn around the whole cell. But anti aliasing is not activated in general. At least on my machine this would make the borders of each grid cell (not highlighted) much thicker and would look inferior to the current version