MainToolBar: hide instead of disable JACK Timebase button
in case the user opted to diable the JACK Timebase feature in the `PreferencesDialog`, there is no real advantange in showing `m_pJackTimebaseButton` disabled. Previously, this was done in order to not break the hand crafted layout of the (former) `PlayerControl`. But in `MainToolBar` we are happy about less items and do hide it instead