PatternFillDialog: full rework
I could stand the dialog just being ugly. But there were also a bug causing Hydrogen to basically freeze when entering an enormouse number in the to dialog.

In addition, the dialog was soo hard to use. Instead of inputs for arbitrary text (?), we do now use integer spin boxes to specify the range. These will adjust their minimum and maximum values in such a way `from` can never be larger than `to`. Also, the maximum number of bars configured in the `PreferencesDialog` is used as upper value for `to`.

Instead of a radio buttons and one Ok button to either fill or clear cell, we now use a "Fill" and "Clear" button to trigger the corresponding action. Also, focus switching using <kdb>Tab</kbd> is now supported